Moved PdoDbContext to root of lib folder. Created outline of db contexts
